requirements.txt
locustio
pytest
pytest-xdist
pytest-html
requests
paramiko
urllib3
rdflib
bs4
wget
pytest-rerunfailures